backend/requirements.txt
flask
flask-cors
pyjwt
python-dotenv
bcrypt
qrcode
pillow
pymongo
